requirements.txt
anyio>=3.7.1,<4.0.0
click==8.1.7
dnspython==2.5.0
fastapi==0.109.2
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
libcst==1.1.0
msgpack==1.0.7
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic>=2.6.0,<3.0.0
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-multipart >=0.0.7
requests>=2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
starlette>=0.36.3,<0.37.0
strawberry-graphql[fastapi]==0.219.2
tomli==2.0.1
typer==0.9.0
typing-extensions>=4.9.0
uvicorn==0.27.1
